refactor: move fork class to domain package
