+++ internal/domain/fork.go
@@ -1,6 +1,6 @@
-package fork
+package domain
 
-type Configuration struct {
+type ForkConfiguration struct {
 	DefaultOrganization string
 	IsInteractive       bool
 }
@@ -18,4 +18,13 @@ type ForkStatus struct {
 	HasCorrectDefault bool
 	ForkName          string
 	UpstreamName      string
+}
+
+type ForkProvider interface {
+	IsRepositoryFork() (bool, error)
+	CreateFork(forkName string) error
+	ForkExists(forkName string) (bool, error)
+	SetDefaultRepository(repo string) error
+	GetRemoteConfiguration() (map[string]string, error)
+	ConfigureRemotesForExistingFork(forkName string) error
 }
